Move `std::io::SizeHint` to `core::io`
